requirements.txt
cleverbot.py
django